project/requirements.txt
minio
pdf2image
torch
transformers
dotenv
qdrant-client
colpali-engine
pillow
